comments

+++ app/Http/Controllers/RentApartmentController.php
@@ -8,6 +8,7 @@ use App\Http\Resources\RentApartment as RentApartmentResources;
 use App\Models\Apartement;
 use App\Models\RentApartment;
 use App\Models\User;
+use Illuminate\Http\Request;
 
 class RentApartmentController extends Controller
 {
@@ -40,31 +41,43 @@ class RentApartmentController extends Controller
     public function store(StoreRentApartmentRequest $request)
     {
         $request->validated();
-        $isExist = RentApartment::where('user_id', '=', $request['user_id'])->where('apartment_id', '=', $request['apartment_id'])->get();
-        if ($isExist->isNotEmpty()) {
-
-            return response([
-                'error' => 'this user already rent this apartment'
-            ], 402);
-        }
+        // 
         $user_gender = User::where('id', $request['user_id'])->first()['gender'];
 
         $apartment = Apartement::where('id', $request['apartment_id'])->first();
 
-        if ($user_gender !== $apartment['gender']) {
+        if ($user_gender !== $apartment['gender']) 
+        {
 
             return response([
                 'data' => 'your gender is not match with requested apartment'
             ], 402);
         }
-        if ($apartment['available'] < 1) {
+        if ($apartment['available'] < 1) 
+            {
             return response([
                 'data' => 'this apartment is full'
             ], 200);
         }
+        // 
+        $isExist = RentApartment::where('user_id', '=', $request['user_id'])->where('apartment_id', '=', $request['apartment_id'])->get();
+        // ->where('status', 'requested')
+        if ($isExist->where('status', 'requested')->first()) {
+
+            return response([
+                'error' => 'this user already request this apartment'
+            ], 402);
+        }
+        /* else if ($isExist->where('status', 'confirmed')->first()) {
+
+            return response([
+                'error' => 'you must leave first, then request'
+            ], 402);
+        } */
+
         RentApartment::create($request->all());
-        Apartement::where('id', $request['apartment_id'])
-            ->decrement('available');
+        /* Apartement::where('id', $request['apartment_id'])
+            ->decrement('available'); */
         return response([
             'data' => 'your request successfully done'
         ], 200);
@@ -128,6 +141,8 @@ class RentApartmentController extends Controller
             // $isExist->update('status', 'confirmed');
             $isExist->status = 'confirmed';
             $isExist->save();
+            Apartement::where('id', $isExist['apartment_id'])
+                ->decrement('available');
             return response([
                 'data' => 'confirming done successfully'
             ], 200);
@@ -155,19 +170,25 @@ class RentApartmentController extends Controller
      * @param  \App\Models\RentApartment  $rentApartment
      * @return \Illuminate\Http\Response
      */
-    public function destroy($rent_id)
+    public function destroy(Request $request, $rent_id)
     {
         $isExist = RentApartment::find($rent_id);
         if ($isExist) {
-            RentApartment::destroy($rent_id);
-            Apartement::where('id', $isExist['apartment_id'])
-                ->increment('available');
-            return response([
-                'data' => 'withdrawing your request successfully'
-            ], 200);
+            if ($isExist['status'] === 'requested') {
+                RentApartment::destroy($rent_id);
+                // Apartement::where('id', $isExist['apartment_id'])
+                //     ->increment('available');
+                return response([
+                    'data' => 'withdrawing / rejecting the request successfully'
+                ], 200);
+            } else {
+                return response([
+                    'error' => 'Can not withdraw / reject confirmed requests'
+                ], 400);
+            }
         }
         return response([
-            'error' => 'this user did not rent this apartment'
+            'error' => 'this rent is not found'
         ], 402);
     }
 }
